requirements.txt
numpy
neopixel
pydevd
Adafruit_MCP3008
Adafruit_GPIO.SPI